requirements.txt
streamlit
tensorflow
numpy
pillow
pandas
requests
